Elephant Quantum Walk as a quantum map of an open system.
Now for every memory element we have a kraus operator, so that the evolution is
described by a operator-sum representation of the EQW map.

The choice of the displacements in the t+1 time step are made accodingly with
 a uniform distribuition in the interval 0-t. With probability p we use the same
displacement in the instant selected and with 1-p we use the opposite.

The class ElephantFermionShiftOperator doenst exists anymore. Maibe we dont need
two function/class to define a ShiftOperator for fermion coins.

Comments arent updated.