Lil sparse arrays, complex data types and general ciclic condition.
The change in the type of sparse array, csr to lil, wasnt loaded in the merge
of the Sparse Trans to master. So now we have lil sparse arrays, with complex
data types.

Also now a change in the ciclic condition for the step operators allows one to
have steps greater than one