gEQW Code
This code is based on the generalized elephant quantum walk, which uses a
q-exponential probability distribuition to select the past instants and
uses themselves as the displacements in a given time.

A trace distance function was included in the statistics file, with some
changes in the main file that still not sufficient.