Trace Distance Subroutine Implemented
Now it is possible to calculate the trace distance time evolution for the given
initial state. A orthogonal state to the given state is used in the calculation
of the trace distance and evolved accordingly with the same evolution operator.

To choose to calculate the trace distance, a boolean parameter was added in the
gEQW.cfg.

A plot of the trace distance x time is generated after the simulation.