gEQW without a full walker state and trace distance derivative
Now the gEQW simulation uses an array of the coin coeficients for every position.
So we dont need the shift operator and the coin operator in all the hilbert spa-
ce. Because of this change, changes were made in the statistics file to calcula-
te the probs dists, variances, entanglement entropy and trace distance properly.

The comments must be updated and the max region considered in the loops also.