Efficient evolution for the initial localized states
Now the evolution runs only over the necessary position coin states.
This was achieved by flattening the array of the lattice nodes coordinates
after selecting the desired region. This only works for 1D and 2D cases.

Also, instead of running over all positions in the position statistics function
now we run only over the necessary postions. To do this we save the necessary
positions in the lattice and update then accordingly with thte possible
position range of the walker.

There is an inefficiency in the gaussian initial distribution if the
walker is locallized in one of the directions.

New comments must be made and a change of the way of selecting the states in
the entanglement function in the statistics file, for more efficiency.