Docstrings updated, efficient coin entropy calculation, simulation.py
The docstrings of the functions in the main, statistics and gEWalks files
were updated, although maybe wrong :s.

In the statistics file, the calculation of the coin entanglement entropy
was made more efficient by considering only the states related to the accessible
positions.

The simulation.py file is a code made to run several simulations with diferent
parameters and for a given set of parameters run several instances for
ensemble purposes. To do that it is utilized the subprocess module with
a wait function to not run all the simulations with diferent parameters at once.