Position probability distribution saved as a .npy file
Now all the position probability distribution is saved, instead of the marginal distributions.
The reason is that in the case where are correlations between the positions degrees we can't
retrieve the joint probability distribution from the marginals.

The main file was changed accordingly to save the probability arrays.

Also, the plot function in the main file was moved to another one that is now imported.