Initial coin state implementation change
The initial coin state implementation was changed. Now it uses the basis state coefficients
instead of the bloch angles. The reason of the change was to include the entangled states
in problems with more than one dimension.

Also, comments were made in the cfg file and corrections in the plots.