Displacement distributions and correlation
The displacements probability distributions were moved to the displacements_distributions.py
file. The gEQWalk.py file was modified accordingly. Now it's possible to specify the
distribution that will be used in the gEQW.cfg file.

A correlation function generator between the steps was also implemented.

The Displacement generator must be change to incorporate the possible call of the correlation function
Maybe there is some comments that must be made.